+++ stockTrainerFrontend/src/components/Graph/Graph.css
