+++ stockTrainerFrontend/src/components/Graph/Graph.jsx
